Latitude and Longitude shrinked to Float i/o Double (database also)

+++ src/main/java/org/oapen/irusuk/dataingestion/Event.java
@@ -17,7 +17,7 @@ public class Event {
 	private String itemId;
 	private String country, countryCode;
 	private String city;
-	private Double latitude, longitude;
+	private Float latitude, longitude;
 	private Integer requests;
 
 	public String getIp() {
@@ -44,11 +44,11 @@ public class Event {
 		return city;
 	}
 
-	public Double getLatitude() {
+	public Float getLatitude() {
 		return latitude;
 	}
 
-	public Double getLongitude() {
+	public Float getLongitude() {
 		return longitude;
 	}
 
@@ -80,11 +80,11 @@ public class Event {
 		this.city = city;
 	}
 
-	public void setLatitude(Double latitude) {
+	public void setLatitude(Float latitude) {
 		this.latitude = latitude;
 	}
 
-	public void setLongitude(Double longitude) {
+	public void setLongitude(Float longitude) {
 		this.longitude = longitude;
 	}
